# contracts/openapi/platform.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Lagerkraft.Platform | platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "AuthApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/choose-tenant": {
      "post": {
        "tags": [
          "AuthApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChooseTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "AuthApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "AuthApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/pin-unlock": {
      "post": {
        "tags": [
          "AuthApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinUnlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/switch-tenant": {
      "post": {
        "tags": [
          "AuthApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchTenantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChooseTenantRequest": {
        "required": [
          "chooser_token",
          "tenant_id"
        ],
        "type": "object",
        "properties": {
          "chooser_token": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "password",
          "totp"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "totp": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LogoutRequest": {
        "required": [
          "refresh_token"
        ],
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "PinUnlockRequest": {
        "required": [
          "device_id",
          "user_id",
          "pin"
        ],
        "type": "object",
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "pin": {
            "type": "string"
          }
        }
      },
      "RefreshRequest": {
        "required": [
          "refresh_token"
        ],
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "SwitchTenantRequest": {
        "required": [
          "tenant_id"
        ],
        "type": "object",
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}